generatedDataset/adxClient/requirements.txt
azure-kusto-data
